requirements.txt
cryptography>=41.0.0
PyYAML>=6.0
fastapi>=0.110.0
Jinja2>=3.0.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.5
python-dotenv>=1.0.0 